Simplify UdpCodec impl for ClientCodec
Make the In type a unit struct, because the return value from the
decode() function isn't currently used: the result of the decode and
decrypt is displayed from within the function. Returning nothing makes
the function very concise. This can be revisited later. Returning a
value would make it possible for the display responsibility to be
in the main part of the program.

* src/access.rs: